+++ src/aead.rs
@@ -1,6 +1,6 @@
 //! Authenticated Encryption with Associated Data Algorithms: AES-GCM, ChaCha20Poly1305
 
-pub use aead::{AeadCore, AeadInPlace, Buffer, Error, NewAead};
+pub use aead::{AeadCore, AeadInPlace, Buffer, Error, KeyInit, KeySizeUser};
 
 #[cfg(feature = "alloc")]
 pub use aead::{Aead, Payload};
@@ -12,27 +12,29 @@ use aead::{
 use ring::aead::{
     Aad, LessSafeKey as Key, Nonce, UnboundKey, AES_128_GCM, AES_256_GCM, CHACHA20_POLY1305,
 };
-use zeroize::Zeroize;
 
 /// Authentication tags
 pub type Tag = GenericArray<u8, U16>;
 
 /// AES-GCM with a 128-bit key
-pub struct Aes128Gcm(GenericArray<u8, U16>);
+pub struct Aes128Gcm(Cipher);
 
 /// AES-GCM with a 256-bit key
-pub struct Aes256Gcm(GenericArray<u8, U32>);
+pub struct Aes256Gcm(Cipher);
 
 /// ChaCha20Poly1305
-pub struct ChaCha20Poly1305(GenericArray<u8, U32>);
+pub struct ChaCha20Poly1305(Cipher);
 
 macro_rules! impl_aead {
     ($cipher:ty, $algorithm:expr, $key_size:ty) => {
-        impl NewAead for $cipher {
+        impl KeySizeUser for $cipher {
             type KeySize = $key_size;
+        }
 
+        impl KeyInit for $cipher {
             fn new(key: &GenericArray<u8, Self::KeySize>) -> Self {
-                Self(*key)
+                let key = UnboundKey::new(&$algorithm, key.as_slice()).unwrap();
+                Self(Cipher::new(key))
             }
         }
 
@@ -49,12 +51,8 @@ macro_rules! impl_aead {
                 associated_data: &[u8],
                 buffer: &mut [u8],
             ) -> Result<Tag, Error> {
-                let key = UnboundKey::new(&$algorithm, self.0.as_slice()).unwrap();
-                Cipher::new(key).encrypt_in_place_detached(
-                    nonce.as_slice(),
-                    associated_data,
-                    buffer,
-                )
+                self.0
+                    .encrypt_in_place_detached(nonce.as_slice(), associated_data, buffer)
             }
 
             fn decrypt_in_place(
@@ -63,8 +61,8 @@ macro_rules! impl_aead {
                 associated_data: &[u8],
                 buffer: &mut dyn Buffer,
             ) -> Result<(), Error> {
-                let key = UnboundKey::new(&$algorithm, self.0.as_slice()).unwrap();
-                Cipher::new(key).decrypt_in_place(nonce.as_slice(), associated_data, buffer)
+                self.0
+                    .decrypt_in_place(nonce.as_slice(), associated_data, buffer)
             }
 
             fn decrypt_in_place_detached(
@@ -75,12 +73,6 @@ macro_rules! impl_aead {
                 _tag: &Tag,
             ) -> Result<(), Error> {
                 unimplemented!(); // ring does not allow us to implement this API
-            }
-        }
-
-        impl Drop for $cipher {
-            fn drop(&mut self) {
-                self.0.zeroize();
             }
         }
     };
@@ -108,7 +100,7 @@ impl Cipher {
     ) -> Result<Tag, Error> {
         self.0
             .seal_in_place_separate_tag(
-                Nonce::try_assume_unique_for_key(nonce).unwrap(),
+                Nonce::try_assume_unique_for_key(nonce).map_err(|_| Error)?,
                 Aad::from(associated_data),
                 buffer,
             )
@@ -125,7 +117,7 @@ impl Cipher {
         let pt_len = self
             .0
             .open_in_place(
-                Nonce::try_assume_unique_for_key(nonce).unwrap(),
+                Nonce::try_assume_unique_for_key(nonce).map_err(|_| Error)?,
                 Aad::from(associated_data),
                 buffer.as_mut(),
             )