+++ src/lib.rs
@@ -1,4 +1,11 @@
+#![no_std]
+#![cfg_attr(docsrs, feature(doc_auto_cfg))]
 #![doc = include_str!("../README.md")]
+#![doc(
+    html_logo_url = "https://raw.githubusercontent.com/RustCrypto/media/7f79a5e/img/ring-compat/logo-sq.png"
+)]
+#![forbid(unsafe_code)]
+#![warn(missing_docs, rust_2018_idioms)]
 
 //! # Features
 //!
@@ -12,27 +19,21 @@
 //!   - `p256`: ECDSA/NIST P-256
 //!   - `p384`: ECDSA/NIST P-384
 
-#![no_std]
-#![cfg_attr(docsrs, feature(doc_cfg))]
-#![doc(
-    html_logo_url = "https://raw.githubusercontent.com/RustCrypto/media/7f79a5e/img/ring-compat/logo-sq.png"
-)]
-#![forbid(unsafe_code)]
-#![warn(missing_docs, rust_2018_idioms)]
-
 #[cfg(feature = "std")]
 extern crate std;
 
 #[cfg(feature = "aead")]
-#[cfg_attr(docsrs, doc(cfg(feature = "aead")))]
 pub mod aead;
 
 #[cfg(feature = "digest")]
-#[cfg_attr(docsrs, doc(cfg(feature = "digest")))]
 pub mod digest;
 
 #[cfg(feature = "signature")]
-#[cfg_attr(docsrs, doc(cfg(feature = "signature")))]
 pub mod signature;
 
 pub use generic_array;
+
+#[cfg(feature = "signature")]
+pub use pkcs8;
+
+pub use ring;