+++ tests/signature/ed25519.rs
@@ -10,31 +10,31 @@ use ring_compat::signature::{
 #[test]
 fn sign_rfc8032_test_vectors() {
     for vector in TEST_VECTORS {
-        let signing_key = SigningKey::from_seed(vector.sk).unwrap();
-        assert_eq!(signing_key.sign(vector.msg).as_ref(), vector.sig);
+        let signing_key = SigningKey::from_slice(vector.sk).unwrap();
+        assert_eq!(signing_key.sign(vector.msg).to_bytes(), vector.sig);
     }
 }
 
 #[test]
 fn verify_rfc8032_test_vectors() {
     for vector in TEST_VECTORS {
-        let verify_key = VerifyingKey::new(vector.pk).unwrap();
-        let sig = Signature::from_bytes(vector.sig).unwrap();
-        assert!(verify_key.verify(vector.msg, &sig).is_ok());
+        let verifying_key = VerifyingKey::from_slice(vector.pk).unwrap();
+        let sig = Signature::try_from(vector.sig).unwrap();
+        assert!(verifying_key.verify(vector.msg, &sig).is_ok());
     }
 }
 
 #[test]
 fn rejects_tweaked_rfc8032_test_vectors() {
     for vector in TEST_VECTORS {
-        let verify_key = VerifyingKey::new(vector.pk).unwrap();
+        let verifying_key = VerifyingKey::from_slice(vector.pk).unwrap();
 
         let mut tweaked_sig = [0u8; Signature::BYTE_SIZE];
         tweaked_sig.copy_from_slice(vector.sig);
         tweaked_sig[0] ^= 0x42;
 
-        let sig = Signature::from_bytes(&tweaked_sig[..]).unwrap();
-        assert!(verify_key.verify(vector.msg, &sig).is_err());
+        let sig = Signature::try_from(&tweaked_sig[..]).unwrap();
+        assert!(verifying_key.verify(vector.msg, &sig).is_err());
     }
 }